move useRestrictedMetrics to related comp
