+++ src/ducks/Watchlists/Widgets/Table/Columns/Toggler/index.js
@@ -1,21 +1,20 @@
 import React, { useMemo, useState, useEffect } from 'react'
 import cx from 'classnames'
 import isEqual from 'lodash.isequal'
-import Button from '@santiment-network/ui/Button'
-import Panel from '@santiment-network/ui/Panel'
 import Icon from '@santiment-network/ui/Icon'
+import Panel from '@santiment-network/ui/Panel'
+import Button from '@santiment-network/ui/Button'
 import Search from '@santiment-network/ui/Search'
 import ContextMenu from '@santiment-network/ui/ContextMenu'
 import Category from './Category'
+import ConfigsMenu from './Configs'
 import { buildColumns, Column } from '../builder'
 import { metrics } from '../../../Filter/dataHub/metrics'
-import { useRestrictedMetrics } from '../../../../gql/hooks'
-import { getCategoryGraph } from '../../../../../Studio/Sidebar/utils'
 import { useTheme } from '../../../../../../stores/ui/theme'
 import { getShadowVars } from '../../../../../../utils/styles'
 import { useUpdateWatchlistTableConfig } from '../gql/mutations'
-import ConfigsMenu from './Configs'
-import { useTableConfig } from '../gql/queries'
+import { useRestrictedMetrics, useTableConfig } from '../gql/queries'
+import { getCategoryGraph } from '../../../../../Studio/Sidebar/utils'
 import styles from './index.module.scss'
 
 const Toggler = ({
